requirements.txt
Django==1.7.7
django-bootstrap3==5.1.1
django-extensions==1.5.1
django-grappelli==2.6.3
django-timezone-field==1.2
pytz==2015.2
six==1.9.0
